# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  address?: string;
  contact?: string;
  email?: string;
  phone?: string;
  createdAt: string;
}

export interface CreateClientRequest {
  name: string;
  address?: string;
  contact?: string;
  email?: string;
  phone?: string;
}

export type EmployeeStatus =
  (typeof EmployeeStatus)[keyof typeof EmployeeStatus];

export const EmployeeStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  role?: string;
  department?: string;
  phone?: string;
  email?: string;
  employeeNumber?: string;
  status: EmployeeStatus;
  createdAt: string;
}

export type CreateEmployeeRequestStatus =
  (typeof CreateEmployeeRequestStatus)[keyof typeof CreateEmployeeRequestStatus];

export const CreateEmployeeRequestStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateEmployeeRequest {
  firstName: string;
  lastName: string;
  role?: string;
  department?: string;
  phone?: string;
  email?: string;
  employeeNumber?: string;
  status?: CreateEmployeeRequestStatus;
}

export type PickupStatus = (typeof PickupStatus)[keyof typeof PickupStatus];

export const PickupStatus = {
  available: "available",
  in_use: "in_use",
  maintenance: "maintenance",
} as const;

export interface Pickup {
  id: number;
  plateNumber: string;
  model?: string;
  brand?: string;
  year?: number;
  capacity?: number;
  status: PickupStatus;
  color?: string;
  createdAt: string;
}

export type CreatePickupRequestStatus =
  (typeof CreatePickupRequestStatus)[keyof typeof CreatePickupRequestStatus];

export const CreatePickupRequestStatus = {
  available: "available",
  in_use: "in_use",
  maintenance: "maintenance",
} as const;

export interface CreatePickupRequest {
  plateNumber: string;
  model?: string;
  brand?: string;
  year?: number;
  capacity?: number;
  status?: CreatePickupRequestStatus;
  color?: string;
}

export type PlanStatus = (typeof PlanStatus)[keyof typeof PlanStatus];

export const PlanStatus = {
  draft: "draft",
  confirmed: "confirmed",
  completed: "completed",
} as const;

export interface Plan {
  id: number;
  name: string;
  clientId: number;
  clientName?: string;
  date: string;
  notes?: string;
  status: PlanStatus;
  createdAt: string;
}

export type PlanDetailStatus =
  (typeof PlanDetailStatus)[keyof typeof PlanDetailStatus];

export const PlanDetailStatus = {
  draft: "draft",
  confirmed: "confirmed",
  completed: "completed",
} as const;

export interface Assignment {
  id: number;
  planId: number;
  employeeId?: number;
  pickupId?: number;
  employee?: Employee;
  pickup?: Pickup;
  position: number;
  notes?: string;
}

export interface PlanDetail {
  id: number;
  name: string;
  clientId: number;
  clientName?: string;
  date: string;
  notes?: string;
  status: PlanDetailStatus;
  createdAt: string;
  assignments: Assignment[];
}

export type CreatePlanRequestStatus =
  (typeof CreatePlanRequestStatus)[keyof typeof CreatePlanRequestStatus];

export const CreatePlanRequestStatus = {
  draft: "draft",
  confirmed: "confirmed",
  completed: "completed",
} as const;

export interface CreatePlanRequest {
  name: string;
  clientId: number;
  date: string;
  notes?: string;
  status?: CreatePlanRequestStatus;
}

export type UpdatePlanRequestStatus =
  (typeof UpdatePlanRequestStatus)[keyof typeof UpdatePlanRequestStatus];

export const UpdatePlanRequestStatus = {
  draft: "draft",
  confirmed: "confirmed",
  completed: "completed",
} as const;

export interface UpdatePlanRequest {
  name?: string;
  clientId?: number;
  date?: string;
  notes?: string;
  status?: UpdatePlanRequestStatus;
}

export type SaveAssignmentsRequestAssignmentsItem = {
  employeeId?: number;
  pickupId?: number;
  position: number;
  notes?: string;
};

export interface SaveAssignmentsRequest {
  assignments: SaveAssignmentsRequestAssignmentsItem[];
}

export interface ImportEmployeesRequest {
  employees: CreateEmployeeRequest[];
}

export interface ImportPickupsRequest {
  pickups: CreatePickupRequest[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  errors: string[];
}