# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from "./employeeStatus";

export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  role?: string;
  department?: string;
  phone?: string;
  email?: string;
  employeeNumber?: string;
  status: EmployeeStatus;
  createdAt: Date;
}