# lib/api-zod/src/generated/types/pickup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
import type { PickupStatus } from "./pickupStatus";

export interface Pickup {
  id: number;
  plateNumber: string;
  model?: string;
  brand?: string;
  year?: number;
  capacity?: number;
  status: PickupStatus;
  color?: string;
  createdAt: Date;
}