# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
import type { PlanStatus } from "./planStatus";

export interface Plan {
  id: number;
  name: string;
  clientId: number;
  clientName?: string;
  date: Date;
  notes?: string;
  status: PlanStatus;
  createdAt: Date;
}