# lib/api-zod/src/generated/types/planDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
import type { Assignment } from "./assignment";
import type { PlanDetailStatus } from "./planDetailStatus";

export interface PlanDetail {
  id: number;
  name: string;
  clientId: number;
  clientName?: string;
  date: Date;
  notes?: string;
  status: PlanDetailStatus;
  createdAt: Date;
  assignments: Assignment[];
}