# lib/api-zod/src/generated/types/saveAssignmentsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
import type { SaveAssignmentsRequestAssignmentsItem } from "./saveAssignmentsRequestAssignmentsItem";

export interface SaveAssignmentsRequest {
  assignments: SaveAssignmentsRequestAssignmentsItem[];
}