# lib/api-zod/src/generated/types/updatePlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Planning App API - Fleet and HR management
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePlanRequestStatus } from "./updatePlanRequestStatus";

export interface UpdatePlanRequest {
  name?: string;
  clientId?: number;
  date?: Date;
  notes?: string;
  status?: UpdatePlanRequestStatus;
}